+++ example/src/shader.rs
@@ -177,8 +177,9 @@ pub const ENTRY_VS_MAIN: &str = "vs_main";
 pub const ENTRY_FS_MAIN: &str = "fs_main";
 #[derive(Debug)]
 pub struct VertexEntry<const N: usize> {
-    entry_point: &'static str,
-    buffers: [wgpu::VertexBufferLayout<'static>; N],
+    pub entry_point: &'static str,
+    pub buffers: [wgpu::VertexBufferLayout<'static>; N],
+    pub constants: std::collections::HashMap<String, f64>,
 }
 pub fn vertex_state<'a, const N: usize>(
     module: &'a wgpu::ShaderModule,
@@ -188,13 +189,50 @@ pub fn vertex_state<'a, const N: usize>(
         module,
         entry_point: entry.entry_point,
         buffers: &entry.buffers,
-        compilation_options: Default::default(),
+        compilation_options: wgpu::PipelineCompilationOptions {
+            constants: &entry.constants,
+            ..Default::default()
+        },
     }
 }
-pub fn vs_main_entry(vertex_input: wgpu::VertexStepMode) -> VertexEntry<1> {
+pub fn vs_main_entry(
+    vertex_input: wgpu::VertexStepMode,
+    overrides: &OverrideConstants,
+) -> VertexEntry<1> {
     VertexEntry {
         entry_point: ENTRY_VS_MAIN,
         buffers: [VertexInput::vertex_buffer_layout(vertex_input)],
+        constants: overrides.constants(),
+    }
+}
+#[derive(Debug)]
+pub struct FragmentEntry<const N: usize> {
+    pub entry_point: &'static str,
+    pub targets: [Option<wgpu::ColorTargetState>; N],
+    pub constants: std::collections::HashMap<String, f64>,
+}
+pub fn fragment_state<'a, const N: usize>(
+    module: &'a wgpu::ShaderModule,
+    entry: &'a FragmentEntry<N>,
+) -> wgpu::FragmentState<'a> {
+    wgpu::FragmentState {
+        module,
+        entry_point: entry.entry_point,
+        targets: &entry.targets,
+        compilation_options: wgpu::PipelineCompilationOptions {
+            constants: &entry.constants,
+            ..Default::default()
+        },
+    }
+}
+pub fn fs_main_entry(
+    targets: [Option<wgpu::ColorTargetState>; 1],
+    overrides: &OverrideConstants,
+) -> FragmentEntry<1> {
+    FragmentEntry {
+        entry_point: ENTRY_FS_MAIN,
+        targets,
+        constants: overrides.constants(),
     }
 }
 pub fn create_shader_module(device: &wgpu::Device) -> wgpu::ShaderModule {